Update README.md
This repository contains a comprehensive analysis of Foreign Direct Investment (FDI) in India from 2000-01 to 2016-17. The dataset includes sector-wise and year-wise FDI inflows, allowing for an in-depth exploration of investment trends and patterns. Key metrics such as total and average investments per sector, year-wise investment trends, and correlation between sectors are analyzed. Additionally, the repository includes time series forecasting to predict future FDI trends. The analysis is conducted using R, with visualizations and code provided to facilitate understanding and reproducibility. This project aims to provide valuable insights for policymakers, investors, and researchers interested in the dynamics of FDI in India.